==> scripts/05_run_linear_models_xpgs.R <==
#01_run_linear_models.R

# This script prepares and runs linear models, testing constraints and the effect
# of sequentially adding covariates

# Load packages

library(tidyverse)
library(lavaan)

# Read in data including xPGS

load( file = './data/processed_data_xpgs.RData')

# Restrict to siblings to avoid overfitting associated with creating the xPGS in the same sample (we used singletons)

sibs <- alldata_new %>% 
  filter(duplicated(m_id))

allsibs <- alldata_new %>% 
  filter(m_id %in% sibs$m_id)


# Source script with model functions

source('./scripts/01a_linear_model_funs.R')

# Create a data frame with inputs and to hold results

linear_mods <- expand.grid(names(allsibs) %>% 
                .[str_detect(.,"cbcl")] %>% 
              str_sub(end=-5) %>% 
              unique() , 
              names(allsibs) %>% 
                .[str_detect(.,"pgs")] %>% 
                str_remove_all(".pgs.pc.child|.pgs.pc.father|.pgs.pc.mother") %>% 
                unique()) %>% 
  `colnames<-`(c("out","mod")) %>% 
  as_tibble()

# Remove some irrelevant combinations (xPGS are outcome specific)


linear_mods <- linear_mods %>% 
  filter(!(str_detect(out, "ext")&str_detect(mod, "int")),
         !(str_detect(out, "int")&str_detect(mod, "ext")))

# Run the models across each combination of outcome and moderator, saving results in nested dataframe

linear_mods_res_xpgs <- linear_mods %>% 
  mutate(results = pmap(., run_test_models, data=allsibs),
         main_effs_tab = map(results, make_main_tab),
         int_effs_tab = map(results, make_int_tab),
         fit_comp_tab = map(results, make_fit_tab))

# Save the result

save(linear_mods_res_xpgs, file="./output/linear_mods_res_xpgs.RData")
